requirements.txt
kdbai-client>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
docker>=6.0.0
faiss-cpu>=1.7.0
qdrant-client>=1.7.0
